신고 화면 리팩터링 1. 위젯 분리 2. 추후 item_add와 공유 위젯 병합 예정
